refactor!: max block number --> include by timestamp (#14980)
Fixes #5757

Changes max block number to be seconds based instead of blocks based.
This results in the property being called `include_by_timestamp` now.
The most tricky thing about this PR was handling genesis state. Having
block number zero in in initial header worked but zero timestamp did
not. It has been decided that to work around this issue the timestamp
check will be disabled for block 1 (and only block 1). This resulted in
a change in test setup where we now wait for the block 1 to by mined.

While timestamp is `u64` we still pack it in 32 bits in SharedMutable.
This is fine until year 2106 so we don't care.

---------

+++ yarn-project/stdlib/src/interfaces/archiver.ts
@@ -65,7 +65,7 @@ export const ArchiverApiSchema: ApiSchemaFor<ArchiverApi> = {
   getBytecodeCommitment: z.function().args(schemas.Fr).returns(schemas.Fr),
   getContract: z
     .function()
-    .args(schemas.AztecAddress, optional(schemas.Integer))
+    .args(schemas.AztecAddress, optional(schemas.BigInt))
     .returns(ContractInstanceWithAddressSchema.optional()),
   getContractClassIds: z.function().args().returns(z.array(schemas.Fr)),
   registerContractFunctionSignatures: z.function().args(z.array(z.string())).returns(z.void()),
